# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Mairie des Jeunes de Covè CMS
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface SiteSettings {
  siteName: string;
  tagline: string;
  logoUrl?: string;
  faviconUrl?: string;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  contactEmail?: string;
  contactPhone?: string;
  contactAddress?: string;
  facebookUrl?: string;
  twitterUrl?: string;
  instagramUrl?: string;
  linkedinUrl?: string;
  youtubeUrl?: string;
  tiktokUrl?: string;
  whatsappNumber?: string;
  seoTitle?: string;
  seoDescription?: string;
  seoKeywords?: string;
  ogImageUrl?: string;
  statYouthCount: number;
  statActivitiesCount: number;
  statProjectsCount: number;
  statPartnersCount: number;
  statYouthLabel?: string;
  statActivitiesLabel?: string;
  statProjectsLabel?: string;
  statPartnersLabel?: string;
  mayorName?: string;
  mayorTitle?: string;
  mayorMessage?: string;
  mayorPhotoUrl?: string;
  footerText?: string;
}

export interface SettingsUpdateInput {
  siteName?: string;
  tagline?: string;
  logoUrl?: string;
  faviconUrl?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  contactEmail?: string;
  contactPhone?: string;
  contactAddress?: string;
  facebookUrl?: string;
  twitterUrl?: string;
  instagramUrl?: string;
  linkedinUrl?: string;
  youtubeUrl?: string;
  tiktokUrl?: string;
  whatsappNumber?: string;
  seoTitle?: string;
  seoDescription?: string;
  seoKeywords?: string;
  ogImageUrl?: string;
  statYouthCount?: number;
  statActivitiesCount?: number;
  statProjectsCount?: number;
  statPartnersCount?: number;
  statYouthLabel?: string;
  statActivitiesLabel?: string;
  statProjectsLabel?: string;
  statPartnersLabel?: string;
  mayorName?: string;
  mayorTitle?: string;
  mayorMessage?: string;
  mayorPhotoUrl?: string;
  footerText?: string;
}

export interface HeroSlide {
  id: number;
  title: string;
  subtitle?: string;
  ctaLabel?: string;
  ctaUrl?: string;
  imageUrl: string;
  sortOrder: number;
  active: boolean;
}

export interface HeroSlideInput {
  title: string;
  subtitle?: string;
  ctaLabel?: string;
  ctaUrl?: string;
  imageUrl: string;
  sortOrder?: number;
  active?: boolean;
}

export interface PageContent {
  id: number;
  slug: string;
  title: string;
  heading?: string;
  subheading?: string;
  body?: string;
  heroImageUrl?: string;
  seoTitle?: string;
  seoDescription?: string;
}

export interface PageContentInput {
  title?: string;
  heading?: string;
  subheading?: string;
  body?: string;
  heroImageUrl?: string;
  seoTitle?: string;
  seoDescription?: string;
}

export interface TeamMember {
  id: number;
  fullName: string;
  role: string;
  bio?: string;
  photoUrl?: string;
  email?: string;
  phone?: string;
  facebookUrl?: string;
  linkedinUrl?: string;
  sortOrder: number;
}

export interface TeamMemberInput {
  fullName: string;
  role: string;
  bio?: string;
  photoUrl?: string;
  email?: string;
  phone?: string;
  facebookUrl?: string;
  linkedinUrl?: string;
  sortOrder?: number;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  planned: "planned",
  ongoing: "ongoing",
  completed: "completed",
} as const;

export interface Project {
  id: number;
  title: string;
  category?: string;
  description?: string;
  body?: string;
  imageUrl?: string;
  status: ProjectStatus;
  startDate?: string;
  endDate?: string;
  location?: string;
  budget?: string;
  progress?: number;
  featured?: boolean;
  createdAt: string;
}

export type ProjectInputStatus =
  (typeof ProjectInputStatus)[keyof typeof ProjectInputStatus];

export const ProjectInputStatus = {
  planned: "planned",
  ongoing: "ongoing",
  completed: "completed",
} as const;

export interface ProjectInput {
  title: string;
  category?: string;
  description?: string;
  body?: string;
  imageUrl?: string;
  status: ProjectInputStatus;
  startDate?: string;
  endDate?: string;
  location?: string;
  budget?: string;
  progress?: number;
  featured?: boolean;
}

export interface NewsArticle {
  id: number;
  slug: string;
  title: string;
  excerpt?: string;
  body?: string;
  coverImageUrl?: string;
  author?: string;
  category?: string;
  published: boolean;
  publishedAt?: string;
  createdAt: string;
}

export interface NewsInput {
  slug?: string;
  title: string;
  excerpt?: string;
  body?: string;
  coverImageUrl?: string;
  author?: string;
  category?: string;
  published?: boolean;
  publishedAt?: string;
}

export interface EventItem {
  id: number;
  title: string;
  description?: string;
  location?: string;
  startsAt: string;
  endsAt?: string;
  imageUrl?: string;
  category?: string;
  featured?: boolean;
}

export interface EventInput {
  title: string;
  description?: string;
  location?: string;
  startsAt: string;
  endsAt?: string;
  imageUrl?: string;
  category?: string;
  featured?: boolean;
}

export interface GalleryPhoto {
  id: number;
  title?: string;
  description?: string;
  imageUrl: string;
  category?: string;
  sortOrder: number;
  createdAt: string;
}

export interface GalleryPhotoInput {
  title?: string;
  description?: string;
  imageUrl: string;
  category?: string;
  sortOrder?: number;
}

export interface Partner {
  id: number;
  name: string;
  logoUrl?: string;
  websiteUrl?: string;
  description?: string;
  sortOrder: number;
}

export interface PartnerInput {
  name: string;
  logoUrl?: string;
  websiteUrl?: string;
  description?: string;
  sortOrder?: number;
}

export interface Opportunity {
  id: number;
  title: string;
  category?: string;
  description?: string;
  body?: string;
  imageUrl?: string;
  deadline?: string;
  applyUrl?: string;
  location?: string;
  organization?: string;
  published: boolean;
  createdAt: string;
}

export interface OpportunityInput {
  title: string;
  category?: string;
  description?: string;
  body?: string;
  imageUrl?: string;
  deadline?: string;
  applyUrl?: string;
  location?: string;
  organization?: string;
  published?: boolean;
}

export interface ContactInput {
  fullName: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
}

export interface ContactMessage {
  id: number;
  fullName: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface DashboardSummary {
  heroSlidesCount: number;
  teamCount: number;
  projectsCount: number;
  newsCount: number;
  eventsCount: number;
  upcomingEventsCount: number;
  galleryCount: number;
  partnersCount: number;
  opportunitiesCount: number;
  unreadMessagesCount: number;
  totalMessagesCount: number;
}