# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Mairie des Jeunes de Covè CMS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentAdmin
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changeAdminPassword
      tags: [auth]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "204":
          description: Password changed
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===== SETTINGS =====
...
  /hero-slides:
    get:
      operationId: listHeroSlides
      tags: [hero]
      summary: List hero slides
      responses:
        "200":
          description: Hero slides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeroSlide"
    post:
      operationId: createHeroSlide
      tags: [hero]
      summary: Create hero slide (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeroSlideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroSlide"
  /hero-slides/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateHeroSlide
      tags: [hero]
      summary: Update hero slide (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeroSlideInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroSlide"
    delete:
      operationId: deleteHeroSlide
      tags: [hero]
      summary: Delete hero slide (admin)
      responses:
        "204":
          description: Deleted

  # ===== PAGES CONTENT =====
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventItem"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventItem"
...
  /events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventItem"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (admin)
      responses:
        "204":
          description: Deleted

  # ===== GALLERY =====
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery photos
      responses:
        "200":
          description: Gallery
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryPhoto"
    post:
      operationId: createGalleryPhoto
      tags: [gallery]
      summary: Add photo (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryPhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
  /gallery/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateGalleryPhoto
      tags: [gallery]
      summary: Update photo (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryPhotoInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
    delete:
      operationId: deleteGalleryPhoto
      tags: [gallery]
      summary: Delete photo (admin)
      responses:
        "204":
          description: Deleted

  # ===== PARTNERS =====
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
...
  /contact/messages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: deleteContactMessage
      tags: [contact]
      summary: Delete message (admin)
      responses:
        "204":
          description: Deleted

  # ===== STATS =====
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required: [id, username]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [currentPassword, newPassword]
...
    HeroSlide:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        subtitle: { type: string }
        ctaLabel: { type: string }
        ctaUrl: { type: string }
        imageUrl: { type: string }
        sortOrder: { type: integer }
        active: { type: boolean }
      required: [id, title, imageUrl, sortOrder, active]
    HeroSlideInput:
      type: object
      properties:
        title: { type: string }
        subtitle: { type: string }
        ctaLabel: { type: string }
        ctaUrl: { type: string }
        imageUrl: { type: string }
        sortOrder: { type: integer }
        active: { type: boolean }
      required: [title, imageUrl]
...
    EventItem:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        location: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        imageUrl: { type: string }
        category: { type: string }
        featured: { type: boolean }
      required: [id, title, startsAt]
    EventInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        imageUrl: { type: string }
        category: { type: string }
        featured: { type: boolean }
      required: [title, startsAt]
    GalleryPhoto:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        category: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, imageUrl, sortOrder, createdAt]
    GalleryPhotoInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        category: { type: string }
        sortOrder: { type: integer }
      required: [imageUrl]
...
    ContactInput:
      type: object
      properties:
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        subject: { type: string }
        message: { type: string }
      required: [fullName, email, subject, message]
    ContactMessage:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        subject: { type: string }
        message: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, fullName, email, subject, message, read, createdAt]
...